+++ src/Classifiers/MultilayerPerceptron.php
@@ -182,7 +182,7 @@ class MultilayerPerceptron implements Estimator, Learner, Online, Probabilistic,
      */
     public function __construct(
         array $hiddenLayers = [],
-        int $batchSize = 100,
+        int $batchSize = 200,
         ?Optimizer $optimizer = null,
         float $alpha = 1e-4,
         int $epochs = 1000,