+++ src/Classifiers/SoftmaxClassifier.php
@@ -132,7 +132,7 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
      * @throws \InvalidArgumentException
      */
     public function __construct(
-        int $batchSize = 100,
+        int $batchSize = 200,
         ?Optimizer $optimizer = null,
         float $alpha = 1e-4,
         int $epochs = 1000,