+++ src/Regressors/MLPRegressor.php
@@ -172,7 +172,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
      */
     public function __construct(
         array $hiddenLayers = [],
-        int $batchSize = 100,
+        int $batchSize = 200,
         ?Optimizer $optimizer = null,
         float $alpha = 1e-4,
         int $epochs = 1000,